Module Zmq.Monitor

type t

type address = string

type error_no = int

type error_text = string

type event =

| Connected of address * Unix.file_descr

| Connect_delayed of address

| Connect_retried of address * int

| Listening of address * Unix.file_descr

| Bind_failed of address * error_no * error_text

| Accepted of address * Unix.file_descr

| Accept_failed of address * error_no * error_text

| Closed of address * Unix.file_descr

| Close_failed of address * error_no * error_text

| Disconnected of address * Unix.file_descr

| Monitor_stopped of address

| Handshake_failed_no_detail of address

| Handshake_succeeded of address

| Handshake_failed_protocol of address * int

| Handshake_failed_auth of address * int

val create : 'a Socket.t -> t

val connect : Context.t -> t -> [< `Monitor ] Socket.t

val recv : ?block:bool -> [< `Monitor ] Socket.t -> event

Receive an event from the monitor socket. block indicates if the call should be blocking or non-blocking. Default true

val string_of_event : event -> string

val mk_string_of_event : unit -> event -> string

Create a memorizing function for converting an event to a string. As its it not possible to reliably retrieve the peer address of a closed socket dues to a race condition, this function pairs connects and disconnects and returns the matching connect peer address to disconnects.

Note that it is not possible to retrieve the peer address of connect events is the peer has disconnected before string_of_event is called