Module Zmq.Msg

type t

type bigstring = (char, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

val init_data : ?⁠offset:int -> ?⁠length:int -> bigstring -> t

Initialize a new message with the given data. The data will be kept alive for the lifetime of the message.

parameter offset

specifies an offset from the start of the given block to use. Defaults to 0.

parameter length

specifies the number of bytes, starting from offset, to use. Defaults the length of data - offset.

val size : t -> int

Size of the message in bytes

val copy_data : t -> bigstring

Retrieve a copy of the data contained in the message.

val unsafe_data : t -> bigstring

Retrieve the data contained in the message.

This is considered unsafe because the underlying data may be freed when the message's lifetime expires.

val close : t -> unit

Free the message. This will be done automatically when the message is garbage collected.

val gets : t -> string -> string

Retrieve a property attached to a message. Property are simple strings. Example of properties include: "Socket-Type", "Identity", "Resource", but underlying transport and security mechanism may add more.