Module Socket.Router

type id_t

Identity of a socket connected to the router.

val id_of_string : string -> id_t

id_of_string s coerces s into an id_t.

val recv : [ `Router ] t -> (id_t * string list) deferred

recv socket waits for a message on socket without blocking other Lwt threads.

val send : [ `Router ] t -> id_t -> string list -> unit deferred

send socket id message sends message on socket to id without blocking other Lwt threads.