Module Zmq_deferred__Socket

module Make : functor (T : Zmq_deferred.Deferred.T) -> sig ... end

This functor is meant to be as compatible as possible with lwt-zmq. It should be straight forward to write a functor over Async_zmq.Socket and Lwt_zmq.Socket.